+++ tls/handshake/pack/src/pack_common.h
@@ -58,7 +58,7 @@ void PackDtlsMsgHeader(HS_MsgType type, uint16_t sequence, uint32_t length, uint
  */
 int32_t PackBufLenError(uint32_t logId, const void *format);
 
-int32_t PackTrustedCAList(HITLS_TrustedCAList *caList, uint8_t *buf, uint32_t bufLen, uint32_t *usedLen);
+int32_t PackTrustedCAList(HITLS_TrustedCAList *caList, uint8_t *buf, uint32_t bufLen, uint16_t *usedLen);
 #ifdef __cplusplus
 }
 #endif /* end __cplusplus */
